Still show feedback while asynchronous effects are finishing
Before, we sould just call stop() on the io_service to cause the
application to stop.  This was not great, because it would cause the
io_service to disregard any remaining events in the queue, and no new
events would be processed.  This means that we would not get any
feedback from the remaining tasts (lets say we are saving a big file
during exit) and potentially chains of effect->update->effect could be
aborted in between.

Now, to quite, we just stop reading from the terminal, and wait until
the io_service runs out of work.  To prevent it from running out of
work too early, we keep a io_service::work object inside the thread of
asynchronous tasks.

The API also has been streamlined, so everything that an effect might
want to know about is kept in a context<Action> object.  For ease of
use, this object is copyable, plus we can convert between them as long
as their Action objects are compatible.  This last aspect is important
for easily composing effects with action subtypes, a pattern we were
already using (see how update() delegates in update_buffer for
applying buffer actions).