Clarify reference semantics of store and context
The rule is: store and context have reference semantics.  Store is
move-only and defines the ownership of the underlying
resources (state, loop, etc.).  The context is copyable but all copies
are "references" that are valid according to their underlying store's
life-time.

Also, in places where we would just take a dispatcher function we now
simply take a context, which can be considered the generic,
type-erased API for "effectful things" to "interact with a store".