Implement pausing the application from the debugger
What we want to achieve is avoid flooding the system while debugging
interactive applications (e.g. a game that ticks every 100th of a
second).

There are several complications here because we are not programming a
real debugger.  The "debugger" [1] state is kept in-process and, even
further, updated synchronously.  So we can not simply "break".  If we
did so, we could not then jump the application back in time, because
the application would not be able to respond to our "jump back" event.

Another option is to simply discard application actions while we are
paused.  However, there actions that probably do really need to be
evaluated in some sequence to avoid entering irrecoverable states.  So
just discarding actions is not an option.

A middle ground is to collaborate between the debugger and the main
loop to achieve the desired behavior:

1. The debugger, while paused, just puts actions in an event queue.
   This queue will be evaluated when resuming, so no action is
   actually lost.

2. The main loop is notified when "paused" or "resumed", so it can
   also decide to stop emitting superfluous actions during that time.

   Potentially, this second mechanism could be used also by the
   application developer to override the definition of a "paused"
   state and add more domain knowledge to it.

All in all, this design is far from perfect and there are alternatives
or other combinations possible.  One option is also to decorate the
loop when enabling the debugger such that asynchronous threads are
blocked and stuff like that.  That could be used blocking on
`dispatch` or even using the `ptrace` API.

Anyways for now we are just experimenting so this is good for now :-)

[1] Maybe we should call it "inspector" or "dev tools", like they do
    in the context of Redux in JS-land.