Add support for passing dependencies as part of the store context
When building the store, we can now use the `with_deps()` enhancer to
add a set of dependencies that are carried inside the context.  These
dependencies are then received by the effects.

The use-case is to make the store aware of the various "services" of
the application required to perform side effects.  For example,
database handles, thead pools, stateful parts of the data model,
caches, etc.  These are passed to the effect functions as part of the
context.

Thanks to how dependencies work (they can be easily subset) an effect
must only declare the dependencies it requires.  Also, multiple
`with_deps` enhancers can be combined, and their dependency sets will
me merged. These two features enable modularity, preventing a global
dependency set.

We can also access the dependencies through the context that we get
from `store::get_context`.  However, we can type erase the
dependencies away (or a subset!) if we do not really care about them
at the point of the store declaration, and would like to defer them to
construction time.  The effects, however, must be aware of all
dependencies that they or their sub-effects can use.