Make the lager::context be a set of dependencies in itself
For this to work properly, we removed the variadic constructor which
intialized with a set of provided dependencies.  Instead, a static
factory method `::with()` is included.  This is more explicit and
removes the need for crazy metaprograming to decide whether we are
converting from another dependency set or providing a set of
dependencies.