Make lager::store a reader cursor
This is a big API change where ther lager::store is made a cursor.
This means we do not need to really provide a view function, instead,
watching the reader cursor provides a more general and versatile API.

In the process, we also decided to get rid of the get_context()
function and make the store itself a context.  This eases the
implementation of the `dispatch()` method and, again, makes everything
a bit more orthogonal and simple maybe.