Fix value shall not be moved causing super weird behavior!
When we do `send_up()` we do a `push_down()` which compares `current_`
with the value that has been passed in.  But current_ is at that time
invalid, because we just moved it!

Now, arguably, maybe this check is not that important at the root
state and we could bypass it.  This solves the issue for now though,
we can decide later when doing a wholistic optimization round.