Make cursors watchers independent of the underlying node
When you watch a cursor, connection will be logically maintained with
whatever underlying cursor it is associated to.

Note that when you change the cursor, the watcher is not emitted.  The
reason is that this may cause problems if watchers are expected to be
called in a specific time of the event loop (e.g. whenever committing
states or other roots is actually done).  Normally whoever assigns the
cursor assignment has all the information available anyways to perform
the notification themselves, if necessary at all.