Add QML based event loop
This event loop ensures that when an action is dispatched in the main
thread, it is processed it is processed before rendering the next
frame.

The easiest way to use this is to expose a QML type that inhertis from
`event_loop_quick_item`.  This can be your root view model, which also
holds the store or a reference to it.