Delete ambiguous assigment operators in `state`
It is possible to assign a state with the new value, which implicitly
creates a new state and replaces the current one. It basically detaches
all the cursors and listeners.

Alternatively, the problem can be solved by making the constructor
explicit, but it may create some inconviniences when using the state.

fixes #135