improve compile time message on assignation failure
Consider the following simple program:

```
int main() {
  lager::reader<int> a;
  lager::reader<std::string> b = a;
  return 0;
}
```

Compiling it will produce about many lines of compiler messages, mostly
relating to template deduction failure of a shared_ptr ctor.

This noisy situation is improved by the use of sfinae statements in the
reader to reader ctor and cursor to reader ctor. In this case, an
attempt to compile the aforementioned program will fail with about 4
lines of compiler messages directly quoting the offending line.

N.B. this commit does not cover the case:

```
lager::reader<int> a = lager::with(...);
```