use rank based traversal to minimize node recompute
The problem with a depth first traversal can arise when a transform node
follows a merge node, and where several of the input of the merge node
change within the same transaction.

In this case, the original depth first send down method would trigger a
recomputation of the transform node having the merge node as an input
several times. But all such recomputation except the last one would be
irrelevant.

When propagating values into the tree, only the full set of change of a
merge node matters to subsequent node.

As a proof of concept, a traversal object is created which can schedule
nodes to be visited as a function of their rank. This traversal object
is such that a node is visited only if all of its parent have been
visited.

This means that the merge node is only evaluated once if several of its
input changes, and any transform node having the merge node as an input
will also only recompute once.