add traversal benchmark
The traversal benchmarks in this commit operates on two type of graph:

1. Simple graph of the form A --> B --> C --> ... in which a parent has
   one child and a child has one parent except for the root and leaf
   node.
2. A "diamond" graph that splits then merge the value back at each
   "link" in the chain.

The second graph is designed to cause extraneous traversal computation
in the DFS strategy, while the first one is used as a kind of baseline.

The expectation is that DFS should perform better on the simple graph
while the topological traversal should perform better in the diamond
graph.