Fix build of http_debug_server with boost >= 1.80
Boost Beast 1.80 switches from boost::string_view to
boost::core::string_view [1][2], and thus the "to_string()" method to
convert to std::string is not available anymore.

Both the old and the new string_view types used support converting to
std::string via operator, hence switch to a static_cast to do the
conversion. This supports building with all the versions of boost.

[1] https://www.boost.org/doc/libs/release/libs/beast/doc/html/beast/release_notes.html
[2] https://github.com/boostorg/beast/issues/2417