revert to index based notify loop
This fixes an issue in which calls to link occurring during notification
would invalidate the iterators of the structured for loop. Calling link
as a result of a reader<T>::watch callback can occur quite naturally.

By using the size of the children vector at the point at which the for
loop is initiated, and by using the [] operator in order to access the
children within the vector, the issue is avoided.